ARTEMIS-2450 page-size-bytes should not be greater than Integer.MAX_VALUE

+++ artemis-server/src/main/java/org/apache/activemq/artemis/core/paging/PagingStore.java
@@ -61,7 +61,7 @@ public interface PagingStore extends ActiveMQComponent, RefCountMessageListener 
 
    long getFirstPage();
 
-   long getPageSizeBytes();
+   int getPageSizeBytes();
 
    long getAddressSize();
